requirements.txt
alembic==1.13.1
blinker==1.7.0
celery==5.4.0
click==8.1.7
cssmin==0.2.0
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Assets==2.1.0
Flask-Caching==2.2.0
Flask-Limiter==3.5.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Talisman==1.1.0
Flask-WTF==1.2.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
jsmin==3.0.1
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.2
Pillow==10.3.0
python-dotenv==1.0.1
python-slugify==8.0.4
PyMySQL==1.1.0
redis==5.0.4
SQLAlchemy==2.0.29
Werkzeug==3.0.2
WTForms==3.1.2
yookassa==2.14.0